chore(modules/brick): move tests to parent folder
